script/1080/pic1080/1/requirements.txt
pyautogui
opencv-python
numpy
pillow
pystray
keyboard
psutil
